+++ crates/common/src/python/actor.rs
@@ -1014,6 +1014,13 @@ impl PyDataActor {
         Ok(())
     }
 
+    #[pyo3(name = "subscribe_order_fills")]
+    #[pyo3(signature = (instrument_id))]
+    fn py_subscribe_order_fills(&mut self, instrument_id: InstrumentId) -> PyResult<()> {
+        self.subscribe_order_fills(instrument_id);
+        Ok(())
+    }
+
     #[cfg(feature = "defi")]
     #[pyo3(name = "subscribe_blocks")]
     #[pyo3(signature = (chain, client_id=None, params=None))]
@@ -1363,6 +1370,13 @@ impl PyDataActor {
         params: Option<IndexMap<String, String>>,
     ) -> PyResult<()> {
         self.unsubscribe_instrument_close(instrument_id, client_id, params);
+        Ok(())
+    }
+
+    #[pyo3(name = "unsubscribe_order_fills")]
+    #[pyo3(signature = (instrument_id))]
+    fn py_unsubscribe_order_fills(&mut self, instrument_id: InstrumentId) -> PyResult<()> {
+        self.unsubscribe_order_fills(instrument_id);
         Ok(())
     }